json working, now trying testing two strings

+++ src/main/java/com/burrsutter/JSONResponse.java
@@ -2,7 +2,7 @@ package com.burrsutter;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 
-public record JSONResponse(String input, String output, int score) {
+public record JSONResponse(String input, String output, float score) {
     @JsonCreator
     public JSONResponse {
